Create WordSplit.java
Have the function WordSplit(strArr) read the array of strings stored in strArr, which will contain 2 elements: the first element will be a sequence of characters, and the second element will be a long string of comma-separated words, in alphabetical order, that represents a dictionary of some arbitrary length. For example: strArr can be: ["hellocat", "apple,bat,cat,goodbye,hello,yellow,why"]. The goal is to determine if the first element in the input can be split into two words, where both words exist in the dictionary that is provided in the second element of the input. In this example, the first element can be split into two words: hello and cat because both of those words are in the dictionary.  The output should be those words separeted by comma: hello,cat
If there is no way to split the string into two words that exist in the dictionary, return the string 'not possible'.
The first element itself will not exist in the dictionary as a real word. 

Examples: 
input: baseball, a,ab,b,ball,bas,base,cat,code,d,e,quit,z
outpu: base,ball

input: abcgefd, a,ab,abc,abcg,b,c,dog,e,efd,zzzz
output: abcg,efd